Extract our sinatra app into its own class so that it can be tested
