+++ lib/machinery.rb
@@ -35,6 +35,8 @@ require "pathname"
 require "nokogiri"
 require "socket"
 require "diffy"
+require "sinatra/base"
+require "mimemagic"
 
 require_relative "machinery_logger"
 require_relative "zypper"
@@ -76,7 +78,6 @@ require_relative "ui"
 require_relative "validate_task"
 require_relative "migration"
 require_relative "upgrade_format_task"
-require_relative "html"
 require_relative "hint"
 require_relative "mountpoints"
 require_relative "config_base"
@@ -99,6 +100,8 @@ require_relative "man_task"
 require_relative "comparison"
 require_relative "serve_html_task"
 require_relative "file_diff"
+require_relative "server"
+require_relative "html"
 
 Dir[File.join(Machinery::ROOT, "plugins", "**", "*.rb")].each { |f| require(f) }
